+++ uniform.h
@@ -11,7 +11,7 @@ class uniform_model : public model
 
     uniform_model(double a, double b, double r)
         : model(a, b, r) {}
-    void generate(unsigned n, unsigned size);
+    void generate(unsigned n, double size);
 };
 
 } // namespace sinr
